+++ include/LightComponent.hpp
@@ -11,11 +11,11 @@ public:
     MColour colour_variation;
     MColour current_gradient;
 
-    float period;
-    float time_to_gradient_change = 0;
-    float range;  // How far does this light shine?
+    double period;
+    double time_to_gradient_change = 0;
+    double range;  // How far does this light shine?
 
-    LightComponent(MColour base_colour, MColour colour_variation, float period, float range)
+    LightComponent(MColour base_colour, MColour colour_variation, double period, double range)
         : Component(LightComponentID),
         base_colour(base_colour),
         current_colour(base_colour),
