+++ include/Vec2.hpp
@@ -5,11 +5,11 @@
 class Vec2 {
 
     public:
-    float x, y;
+    double x, y;
 
     Vec2() : x(0), y(0) {};
-    Vec2(float i) : x(i), y(i) {}
-    Vec2(float i, float j) : x(i), y(j) {}
+    Vec2(double i) : x(i), y(i) {}
+    Vec2(double i, double j) : x(i), y(j) {}
 
     Vec2& operator += (const Vec2 & v){
         x += v.x;
@@ -55,7 +55,7 @@ class Vec2 {
         return t;
     }
 
-    static float dist_sq(const Vec2 &a, const Vec2 &b){
+    static double dist_sq(const Vec2 &a, const Vec2 &b){
         return (a.x - b.x)*(a.x - b.x) + (a.y - b.y)*(a.y - b.y);
     }
