+++ src/Renderer.cpp
@@ -23,7 +23,7 @@ std::vector<SDL_Event> Renderer::get_input_events(){
     return events;
 }
 
-void Renderer::renderTextureToScreen(SDL_Texture *texture, float x, float y, float width, float height){
+void Renderer::renderTextureToScreen(SDL_Texture *texture, double x, double y, double width, double height){
     SDL_Rect screen_location;
     screen_location.x = x - width / 2.0;
     screen_location.y = y - height / 2.0;
